requirements.txt
# Backend dependencies for PassPrint
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
stripe==8.0.0
qrcode[pil]==7.4.2
Pillow==10.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.1.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
boto3==1.34.34
python-magic==0.4.27
reportlab==4.0.7
weasyprint==61.2